+++ src/time/localtime_r.c
@@ -3,8 +3,8 @@
 #include <los/time.h>
 
 struct tm* localtime_r(const time_t* timep, struct tm* buf) {
-    int64_t time_zone = get_time_zone();
-    int64_t dst = time_zone & 1;
+    isize time_zone = get_time_zone();
+    isize dst = time_zone & 1;
     time_zone = time_zone & ~1;
 
     time_t gmt_time = *timep + time_zone;
